requirements.txt
django == 4.2
djangorestframework == 3.16
djangorestframework-simplejwt == 5.5.1
drf-spectacular == 0.28.0